change quantity type from integer to decimal

+++ src/Smartstore.Core/Catalog/Attributes/Domain/ProductVariantAttributeCombination.cs
@@ -181,7 +181,7 @@ namespace Smartstore.Core.Catalog.Attributes
         /// <summary>
         /// Gets or sets the stock quantity.
         /// </summary>
-        public int StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether to allow orders when out of stock.
