+++ src/Smartstore.Core/Checkout/Orders/Reporting/BestsellersReportLine.cs
@@ -18,6 +18,6 @@
         /// <summary>
         /// Gets or sets the total quantity.
         /// </summary>
-        public int TotalQuantity { get; set; }
+        public decimal TotalQuantity { get; set; }
     }
 }
